fix(widgets): added back plugin comments to widget files
Transposh widgets are loaded by the following line

open-transposh/src/Widgets/Plugin_Widget.php:399

This line calls the WordPress function get_plugin_data, this expects there to be a plugin comment block in the php files. It looks like these were removed in the refactor, this commit simply adds the comments back as they were in the original repository.